PROOFREADING/PROOFREADING_WEB/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
scikit-image>=0.21.0
numpy>=1.24.0
Werkzeug>=3.0.0